🗂️ Separates enpoints tests into different files
